making some usefull chenges
